Split WASM loader responsibilities
